move firstVendor conversation functions to utils
