+++ client/utils.js
@@ -1,6 +1,8 @@
 import stringSimilarity from 'string-similarity'
 
 /* global webkitSpeechRecognition webkitSpeechGrammarList webkitSpeechRecognitionEvent */
+
+/* Speech Utils */
 let SpeechRecognition
 let SpeechGrammarList
 let SpeechRecognitionEvent
@@ -53,4 +55,64 @@ export const checkAnswer = (userInput, responses) => {
     }
   }
   return null
+}
+
+/* Conversation Utils */
+export const converse = function() {
+  this.listenToUser = listenToUser.bind(this)
+  this.grade = grade.bind(this)
+  let currentPrompt = this.props.currentPrompt
+  if(!Object.keys(currentPrompt).length) {
+    let firstPrompt = this.props.prompts.find((prompt) => {
+      return prompt.id === this.props.firstPromptId
+    })
+    currentPrompt = firstPrompt
+    this.props.setCurrentPrompt(currentPrompt)
+  }
+  this.listenToUser(currentPrompt)
+  .then((speech) => {
+    let result = this.grade(speech)
+    if (result) {
+      this.setState({
+        incorrectCount: 0,
+        hintText: `You said: ${result.text}`
+      })
+      let nextPrompt = this.props.prompts.find((prompt) => {
+        return prompt.id === result.prompt_responses.nextPromptId
+      })
+      if(nextPrompt) {
+        this.props.setCurrentPrompt(nextPrompt)
+        this.converse()
+      } else {
+        reward()
+        this.props.setCurrentPrompt({})
+        this.setState({hintText: ''})
+      }
+      this.props.clearResponse()
+    } else {
+      this.setState({incorrectCount: this.state.incorrectCount + 1})
+      if(this.state.incorrectCount > 1) {
+        this.setState({hintText: `The vendor said: ${currentPrompt.text}`})
+      }
+      this.props.getVendorResponse(this.state.vendorResponse, this.props.language.nativeLang, this.props.language.learningLang)
+      this.setState({vendorResponse: this.props.vendorResponse})
+      this.converse()
+    }
+  })
+}
+
+function listenToUser(currentPrompt) {
+  return this.props.listen(speechRecObject, {
+    responses: currentPrompt.responses,
+    language: this.props.language
+  })
+}
+
+function grade(answer) {
+  return this.props.checkAnswer(answer, this.props.currentPrompt.responses)
+}
+
+function reward() {
+  //temporary log until we have the rest of the logic for this down
+  console.log('good job duderino, you did the thing!')
 }
